fix: form bundle attribute extension
